update latest source

+++ arch/m68k/mac/macboing.c
@@ -114,8 +114,7 @@ static void mac_init_asc( void )
 			 *   16-bit I/O functionality.  The PowerBook 500 series computers
 			 *   support 16-bit stereo output, but only mono input."
 			 *
-			 *   Technical Information Library (TIL) article number 16405. 
-			 *   http://support.apple.com/kb/TA32601 
+			 *   http://til.info.apple.com/techinfo.nsf/artnum/n16405
 			 *
 			 * --David Kilzer
 			 */
@@ -163,7 +162,7 @@ static void mac_init_asc( void )
 void mac_mksound( unsigned int freq, unsigned int length )
 {
 	__u32 cfreq = ( freq << 5 ) / 468;
-	unsigned long flags;
+	__u32 flags;
 	int i;
 
 	if ( mac_special_bell == NULL )
@@ -225,7 +224,7 @@ static void mac_nosound( unsigned long ignored )
  */
 static void mac_quadra_start_bell( unsigned int freq, unsigned int length, unsigned int volume )
 {
-	unsigned long flags;
+	__u32 flags;
 
 	/* if the bell is already ringing, ring longer */
 	if ( mac_bell_duration > 0 )
@@ -272,7 +271,7 @@ static void mac_quadra_start_bell( unsigned int freq, unsigned int length, unsig
 static void mac_quadra_ring_bell( unsigned long ignored )
 {
 	int	i, count = mac_asc_samplespersec / HZ;
-	unsigned long flags;
+	__u32 flags;
 
 	/*
 	 * we neither want a sound buffer overflow nor underflow, so we need to match
